30881 Feat: Move gordo client code
