+++ gordo/client/schemas.py
@@ -3,50 +3,6 @@ from typing import Any, Dict, Optional
 
 from pydantic import BaseModel, Field
 
-machine = {
-    "name": "test-model",
-    "dataset": {
-        "target_tag_list": ["TRC1", "TRC2"],
-        "data_provider": {"min_size": 100, "max_size": 300, "type": "RandomDataProvider"},
-        "resolution": "10T",
-        "row_filter": "",
-        "known_filter_periods": [],
-        "aggregation_methods": "mean",
-        "row_filter_buffer_size": 0,
-        "asset": None,
-        "default_asset": None,
-        "n_samples_threshold": 0,
-        "low_threshold": -1000,
-        "high_threshold": 50000,
-        "interpolation_method": "linear_interpolation",
-        "interpolation_limit": "8H",
-        "filter_periods": {},
-        "tag_normalizer": "default",
-        "train_start_date": "2015-01-01T00:00:00+00:00",
-        "train_end_date": "2015-06-01T00:00:00+00:00",
-        "tag_list": ["TRC1", "TRC2"],
-        "type": "RandomDataset",
-    },
-    "model": {"sklearn.decomposition.PCA": {"svd_solver": "auto"}},
-    "metadata": {
-        "user_defined": {},
-        "build_metadata": {
-            "model": {
-                "model_offset": 0,
-                "model_creation_date": None,
-                "model_builder_version": "1.1.0",
-                "cross_validation": {"scores": {}, "cv_duration_sec": None, "splits": {}},
-                "model_training_duration_sec": None,
-                "model_meta": {},
-            },
-            "dataset": {"query_duration_sec": None, "dataset_meta": {}},
-        },
-    },
-    "runtime": {"reporters": []},
-    "project_name": "project-name",
-    "evaluation": {"cv_mode": "full_build"},
-}
-
 
 class CrossValidationMetaData(BaseModel):
     scores: Dict[str, Any] = Field(default_factory=dict)
@@ -86,9 +42,8 @@ class Machine(BaseModel):
     dataset: Dict[str, Any] = Field(...)
     metadata: Optional[Metadata] = Field(default_factory=Metadata)
     runtime: Dict[str, Any] = None
-    evaluation: Optional[Dict[str, Any]] = Field(default_factory=dict(cv_mode="full_build"))
+    evaluation: Optional[Dict[str, Any]] = Field(default=dict(cv_mode="full_build"))
 
     def __init__(self, **data):
         super().__init__(**data)
         self.host = f"gordoserver-{self.project_name}-{self.name}"
-
